Moves database models into a persistence namespace
