+++ lib/reporter.js
@@ -3,7 +3,7 @@
 var async = require('async');
 var parse = require('url-parse');
 var request = require('request');
-var models = require('../models');
+var models = require('../models/persistence');
 var ReportSum = require('./report_sum');
 var store_calculations = require('./store_calculations');
