+++ lib/store_calculations.js
@@ -5,7 +5,7 @@ const parse     = require('url-parse');
 const request   = require('request');
 
 const ReportSum = require('./report_sum');
-const models    = require('../models');
+const models    = require('../models/persistence');
 
 module.exports = function store_calculations(githubData, travisData, done) {
   done = done || function () {};
