+++ routes/commit.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var models = require('../models');
+var models = require('../models/persistence');
 
 exports.get = function (req, res){
     var reqCommit = req.params.commit.replace('.html','');
